Move the `noqa` module into `comments`
We are about to make this less specific to the `noqa` kind of comment so
we can use it for `fixit`, too.
